code/src/requirements.txt
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
selenium>=4.0.0
pytest>=7.0.0
pytest-html>=4.1.0
pytest-cov>=4.1.0
python-docx>=0.8.11
PyPDF2>=3.0.0
chardet>=5.0.0
jsonschema>=4.0.0
rich>=13.0.0  # For better console output
typing-extensions>=4.0.0
docker>=6.1.2
jinja2==3.1.2
PyYAML>=6.0.1  # Added for YAML file handling